requirements.txt
backports.weakref==1.0.post1
bleach==1.5.0
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
html5lib==0.9999999
idna==2.6
Markdown==2.6.9
mock==2.0.0
mpmath==1.0.0
nltk==3.2.5
numpy==1.13.3
pbr==3.1.1
pkg-resources==0.0.0
protobuf==3.5.0.post1
requests==2.18.4
sacrebleu==1.1.7
six==1.11.0
sympy==1.1.1
tensor2tensor==1.2.6
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
typing==3.6.2
urllib3==1.22
Werkzeug==0.12.2
